// src/language-server/models/EndurantType.ts
export enum EndurantTypes {
  TYPE = "type",
  HISTORICAL_ROLE = "historicalRole",
  HISTORICAL_ROLE_MIXIN = "historicalRoleMixin",
  EVENT = "event",
  SITUATION = "situation",
  CATEGORY = "category",
  MIXIN = "mixin",
  ROLE_MIXIN = "roleMixin",
  PHASE_MIXIN = "phaseMixin",
  KIND = "kind",
  COLLECTIVE = "collective",
  QUANTITY = "quantity",
  RELATOR = "relator",
  QUALITY = "quality",
  MODE = "mode",
  INTRINSIC_MODE = "intrinsicMode",
  EXTRINSIC_MODE = "extrinsicMode",
  SUBKIND = "subkind",
  ROLE = "role",
  PHASE = "phase",
  ENUMERATION = "enumeration",
  DATATYPE = "datatype",
  ABSTRACT = "abstract",
}